Moved exceptions into common package
