+++ packages/common/README.md
@@ -1,0 +1,9 @@
+# MDIP common
+
+Common MDIP utilities and errors
+
+## Installation
+
+```bash
+npm install @mdip/common
+```
