+++ packages/common/src/exceptions.js
