--- packages/exceptions/README.md
@@ -1,9 +1,0 @@
-# MDIP exceptions
-
-Standard MDIP exception strings
-
-## Installation
-
-```bash
-npm install @mdip/exceptions
-```
