+++ tests/keymaster.test.js
@@ -6,7 +6,7 @@ import * as gatekeeper from '@mdip/gatekeeper/lib';
 import * as cipher from '@mdip/cipher/node';
 import * as db_json from '@mdip/gatekeeper/db/json';
 import * as wallet from '@mdip/keymaster/db/json';
-import { InvalidDIDError, ExpectedExceptionError, UnknownIDError } from '@mdip/exceptions';
+import { InvalidDIDError, ExpectedExceptionError, UnknownIDError } from '@mdip/common/exceptions';
 
 beforeEach(async () => {
     await db_json.start('mdip');
